==> cypher_generator/1_KG_Data_Extractor.ipynb <==
# Knowledge Graph Data Extraction

## Description

In this notebook:  

- connect to an existing Neo4j graph database,
- extract schema and relevant KG information,
- extract a set of instances to use as samples,
- save relevant files for later use.

## Workspace Setup

%pip install neo4j
%pip install python-levenshtein

# Load and mount the drive helper
from google.colab import drive

# This will prompt for authorization
drive.mount('/content/drive')

# Set the working directory
%cd '/content/drive/MyDrive/cypherGen/'

# Upload the Neo4j graph database credentials
# Neo4j graph database credentials
#URI = 'neo4j+s://xxxxxxxx.databases.neo4j.io'
#USER = 'neo4j'
#PWD = 'your password here'

from google.colab import userdata
URI = userdata.get('URI')
PWD = userdata.get('PWD')

# Necessary imports
import neo4j
import pandas as pd
import random
import itertools

# Import the local modules
from utils.utilities import *
from utils.neo4j_conn import *
from utils.neo4j_schema import *
from utils.graph_utils import *

# Initialize the Neo4j connector and utilities modules
graph=Neo4jGraph(url=URI, username='neo4j', password=PWD)
gutils = Neo4jSchema(url=URI, username='neo4j', password=PWD)

# Check the graph connection
graph.query("MATCH (n) RETURN count(n)")

# Create a path variable for the data folder
data_path = '/content/drive/MyDrive/cypherGen/datas/'

# Set file names
schema_file = 'schema_file.json'  # schema as JSON object
formatted_schema_file = 'formatted_schema.txt' # schema as a string to be included with prompt
node_instances_file = 'node_instances_file.json' # set of node instances as JSON object
rels_instances_file = 'rels_instances_file.json' # set of relationship instances as JSON object

## Extract data from KG

# Build the string schema
schema = gutils.get_schema

# Save the string schema to a file
with open(data_path+formatted_schema_file, 'w') as f:
    f.write(schema)

# The string schema
print(schema)

# Build the json schema
jschema = gutils.get_structured_schema
# Check the output
jschema.keys()

# Extract the list of nodes
nodes = get_nodes_list(jschema)
print(nodes)

# Read the nodes with their properties and their datatypes
node_props_types = jschema['node_props']
# Check the output

print(f"The properties of the node Report are:\n{node_props_types['Report']}")

# Extract the relationships
relationships = jschema['relationships']
print("The relationships in the graph are:\n")
relationships

# Extract node samples from the graph - 4 sets of node samples
node_instances = gutils.extract_node_instances(nodes, # list of nodes to extract labels
                                               4)  # how many instances to extract
# We have a list of sublists, one for each node label in the provided list
node_instances[2]

# Extract relationship instances
rels_instances = gutils.extract_multiple_relationships_instances(relationships, # list of relationships to extract instances for
                                                                 8)  # how many instances to extract for each relationship
# A list of sublists with 8 entries for each relatonship type
rels_instances[0][0]

# Serialize extracted neo4j.time data - for saving to json files
nodes_instances_serialized = serialize_nodes_data(node_instances)
rels_instances_serialized = serialize_relationships_data(rels_instances)

# When working with large schema KG is better to provide a subschema only, based on a node selection
# This will extract first neighbors and all the corresponding relationships
get_subgraph_schema(jschema, ['Topic'], # nodes to extract information for
                    2, # Levenshtein distance (actual node label, provided label)
                    True) # formated as a string

# Find the datatypes present in the graph
dtypes = retrieve_datatypes(jschema)
dtypes

# Save data to json files
write_json(jschema, data_path+schema_file)
write_json(nodes_instances_serialized, data_path+node_instances_file)
write_json(rels_instances_serialized, data_path+rels_instances_file)
